Update project to use ring-jetty9-adapter for websockets
I am in the research phase, trying to find out the best way to allow the server
to send push notifications to the client. I spent a decent amount of time
fighting to just use a ring :configurator, but saw that other experienced
this and there was no solution. I decided to go with jetty9 because the
sample project was easy to use as a basis to provide my current project
with the websocket functionality.

web-whiteboard.prototyping.websocktes.simple-server was added to demonstrate
the websocket functionality provided by the adapter. I wanted to poke around
and have that available for testing out new ideas.

lein-ws-demo script was created to give a quick way to run the simple-server.

web-whiteboard.web is the namespace created for what will be the http/websocket
server. It defines a -main so that *lein run* will launch it.