Update websocket prototype to use core.async
core.async was brought in to help manage the events that the websocket
will create. Also worth mentioning is that I added a 10s default timeout
for the websocket connection to make it so that my server and client don't
start out with the assumption that they will remain connected forever.

I put in some code to handle creating a new websocket once it is detected
that the server has closed the current one.