Update prototype to broadcast :pen-move events to connected clients
I changed simple_server.clj quite a bit here, most notably removing
the example accept/reject functions that were used to illustrate how
to route to many different paths for websockets. At this point, the
file is getting a little too noisy to keep all of it around. I pulled
app-state together to allow the server to check for connected clients
and the pen-move-handler pushes events to everyone connected except
the original sender.

Given this progress, I should be able to bring the v002 demo up to
mirroring two open browsers with each of their events.