Add web-whiteboard.client as home for the web client
I was able to get the very rough first pass done here.
I still haven't seen updates from one browser to another, but I am
getting really close.

The client is split into a few areas #{handlers.websocket, state, and ui}.
The only thing I was not fully happy with was that creating the app-state
needed to be done in two steps, because of the way that javascript creates
WebSockets.