Add websocket message support for multiple connected clients (without drawing)
web-whiteboard.client.state

Added get-query-params, which may be more appropriate in carafe,
but for now I wanted to be able to pass client-id and whiteboard-id
through query parameters.

web-whiteboard.client.ui

Added margin to correct for the event position (due to margin of body).
I may want to change this in the future, but for now I wanted to be able
to click and have the svg circle draw where the cursor point is.

Also, the pen-event-handler was changed to have an override. This was
needed because when a mousedown happens, there needs to be an event
that draws for the single click. There was a race condition wrt to
accessing the app-state in time, so I just ensured that it could be
controlled directly.

web-whiteboard.server.handlers.websockets

Added a try/catch for the jetty/send! to make sure that clients that are
no longer connected don't hose the server