Create a UI channel to process actions
The idea I want to use here is for the workflows of the whiteboard, I
want a websocket generated event to produce an action that is equivalent
to the user doing it locally. Using a channel should allow me to make it
so that the UI doesn't care where the request is from.