Add UI support for multiple connected clients
web-whiteboard.client.handlers.websocket

Changed location in app-state of the channel, so updates were just
to make that consistent.

web-whiteboard.client.state

Pulled configure-chans logic into init-app-state so that there aren't
intermediate steps during start up. Changed the structure of app-state
for :channels to be less confusing.

web-whiteboard.client.ui

Restructured the way that pen-event-handler works, it now puts a message
on the ui-chan. The whole point of this is that all events to update the
ui can just go onto the ui-chan and this unifies the interface for when
the user is drawing, or when the client receives a message from the
websocket server. Both just put actions onto the ui-chan.