Refactor :circle drawing mode into it's own home
user

Updated the user namespace to pull in the actual server instead of the prototype.

log4j

Updated the log to point to the websocket namespace, it was useful here to debug
the multiple connected client issues I was experiencing.

client.handlers.websocket

Updated to re-register when client needs to reconnect to the websocket server

state

Added drawing-algorithm state to use an event-handler, draw-handler, and state.
More on that below...

client.ui

Updated the ui to use event-handler and draw-handler from app-state.
This enables more drawing modes to be defined, and all they need to do is
manage their own state and transform the mouse events into the data and state
that each needs.

server.handlers.websocket

Updated the server to handle :pen-move, :pen-up, and :pen-down messages. Also
update the server to handle reconnection more gracefully. It will not try to
send to close websockets anymore and it will replace the websocket references
it has with new register messages.