Add line drawing mode
server.handlers.websockets

Updated the names of bindings so that it is clear that the websocket client
and server communicate using *transit-msg*, and each has to decode that into
a *msg*.

I gave some thought to the idea of passing ui-actions over the wire instead
of the events themselves. I was grappling with the idea of what was more
important. Should users be able to pick a mode and have that mode carried with
their data, or should the data be interpreted based on the mode that a user
has selected. I went with the second one, but this is an area that further
thought could go into in order to see what consequences that would have.

web-whiteboard.client.draw.line

This is where the mode is defined, the first to maintain some of it's own state.
It is worth mentioning that in order to allow the user to click to draw a point,
even if not creating a true line, I handle this case by first drawing a line and
then creating an svg path element. The id of this element is kept for state so that
when :pen-move events come in after a :pen-down, they will update the *d* attribute
to add to the path. A :pen-up finishes the sequence (although is not really necessary
at this time). I want to add some of the optimizations discussed by [Jack Schaedler] (https://jackschaedler.github.io/handwriting-recognition/).Add line drawing mode