Untangle client websocket connection logic
Remove register-client from run

Update create-ws to include ws parameter in callbacks

Update reconnect-ws to not swap! in the ws, since open-handler does that work now

Update open-handler to close existing ws connection, swap! in the new ws, and call
register-client. All of this was done to support the fact that creating a new
websocket made the app-state and reconnect logic more complicated, and there was
actually an issue wrt the server getting bombarded with requests until this version.

Remove swap! from create-app-state since it is no longer necessary