Update client to use a channel for sending websocket messages
I had already implemented a channel for messages received by the
websocket server, which made it easy for the ui and websocket
handler to pay attention to anything from the server. This change
now takes that approach for sending messages to the websocket server.
A :to channel was set up, and the client puts messages on it, instead
of using the websocket handler's send method directly.

This makes the system a little more symmetric in terms of design, the
client uses channels to request/respond from/to the websocket connection.