Add pointer-events css to prevent weird errors
When I would draw many lines close together, there were times when
a whiteboard icon would show, as if I had clicked on one of the
existing strokes and that was causing the browser to respond to an
unknown action. By adding the pointer-events: none, I no longer see
these events.