Add undo/redo functionality
-client.ui

Cleaned up some of the channel code to make it more clear.

Refactored the keyboard bindings to remove common code into
publish-ui-action-wrapper. Created :undo-stroke and :redo-stroke
bindings for undo and redo.

Added update-history-state function to handle capturing history data to
implement undo/redo.

-client.state

Added HistroyMap to the app-state so that undo/redo data can be stored

-client.draw.line and -client.draw.smooth-line

Updated pendown to create the svg circle with a name based on the path-id
so that the undo can find and remove both elements.

Added :undo-stroke to draw-handler to support undo

-server.handlers.websocket

Updated to support the :undo-stroke and :redo-stroke events.

-other
Minor edits to doc and data definition