Refactor keybinding code and add number bindings to colors
This commit took the keybinding code that used to be in ui.cljs
and moved it into it's own namespace. This serves a couple of purposes,
the most important one was to organize and give good names to the idea
of keyboard shortcuts.

I landed on calling a set of keys a 'key-binding', and the main idea
now is that each key-binding either has a handler or it doesn't. Single
as well as multiple key combinations are now considered. The previous
implementation chose to use single keys, but I realize that in order
to expand functionality in the future, it is better to view the keys
in a set.

Moving the keybinding code out makes both ui.cljs and keybinding.cljs
easier to think about, and I am pretty pleased with how short most
of the functions are. It should be easy now to extend things.