Correctly set value for color-picker
There was a bug where using the color-picker input would make it so
that the keybindings to change the color no longer worked.

This was because the keybindings were setting the value attribute of
the color picker element, but selecting a color using the input directly
updates the value property of the element. The issue was that when the
value property is set, the value attribute is shadowed.

The fix required setting the value property instead of the value attribute.
Thank you, JavaScript.